xSDKTrilinos: Fixed some more build errors reported by Ross
The MueLu example couldn't possibly have built before.  I think I
broke it when I was trying to simplify the examples.  Things build
correctly against the master branch of Trilinos, which is good
enough for now, but I do need to revisit that example later to
make sure it's doing what it should.