Allow xSDKTrilinos to build as own CMake project
This is bit of a hack but, for the most part, this was pretty easy.  Here is
how I did this:

1) Added the necessary TriBITS project files ProjectName.cmake,
PackagesList.cmake, TPLsList.cmake and Version.cmake.

2) Created a new TPL called TrilinosTpl and made xSDKTrilinos only have a
required dependence on this TPL, in addition to PETSc and HYPRE.

3) The FindTPLTrilinosTpl.cmake file just calls FIND_PACKAGE(Trilinos).  There
is a slight hack relating to the need to call LINK_DIRECTORIES().  This is
really unfortunate.  Hopefully the <Package>Config.cmake files genreated by
TriBITS don't require this hack.  We will address that as pat of
TriBITSPub/TriBITS#63.

4) Had to copy in Teuchos_StandardParallelUnitTestMain.cpp because Trilinos
does not install this file.

5) Had to enabled variables like ${PACKAGE_NAME}_ENABLE_MueLu since these
packages don't really exist in the stand-alone build of xSDKTrilinos (at least
as TriBITS is concerned).

When I get around to the major refactoring and enhancement
TriBITSPub/TriBITS#63 then these hacks can be removed and the build of
xSDKTrilinos will be much cleaner.