Add xSDKTrilinos_config.hpp for HAVE_XXX_EPETRA, run all examples (SDK-79)
With this commit, now xSDKTrilinos confgiures an xSDKTrilinos_config.hpp file
that will define or not define HAVE_XSDKTRIINOS_EPETRA depending if Epetra is
enabled or not.

Now we can run all of the examples with CTest and they "pass" (i.e. they don't
segfault and return 0).

Next the examples will be made to check pass/fail internally so that we esure
that these examples really do work.