Add -lX11 and -lssl for linking with PETSc (#2)
I hacked and added -lX11 and -lssl to xSDKTrilinos_EXTRA_LINK_FLAGS since
PETSc (or one of the packages that PETSc is building) is finding and linking
against these libraries on the ORNL CASL Fissile4 machines.

NOTE: This is not really a hack since this script is specific to the Fissile4
machines anyway.