Read compilers, compile flags, etc from TrilinosConfig.cmake (xSDKTrilinos #2)
Now xSDKTrilinos when configured as its own CMake project will automatically
use the compilers, compiler flags, linker, etc, that was used for Trilinos, as
read from the TrilinosConfig.cmake file.

Therefore, you can now have a universial do-configure script which is agnostic
of what system you are actually on.

I put in a hack to the Fissile4 Trilinos configure script to add -lXll and
-lssl so that this works with PETSc on the Fissile4.  This is a bit of a hack
but who cares?