xSDKTrilinos: Removed unwise error check from failing examples
Several of the linear solver tests were checking the error
| x - x_k | as well as the residual | b - A x_k |.  If the residual
is more than the tolerance, that's bad and the test SHOULD fail.
The failing test reported by Barry Smith et al at ANL was checking
whether the error was less than some constant value, which was never
a good idea.  We could theoretically check whether
|x - x_k| / |x_k| <= cond(A) |r| / (|A| |x_k|)
but I'm just going to disable the error check altogether; I don't want
to introduce any messiness involving the condition number of A to a
straightforward example.

Many thanks to the ANL folks for reporting this bug as issue #5 on
github!